requirements.txt
# Core dependencies
gitpython==3.1.40
pydantic==2.4.2
pytest==7.4.3
python-dotenv==1.0.0
pytest-asyncio==0.21.1

# Analysis tools
radon==6.0.1
pylint==3.0.2
mypy==1.7.0

# AI/ML frameworks
tensorflow-hub==0.15.0
torch>=2.2.0
transformers==4.35.2
openai>=1.12.0

# Solana specific
solana==0.30.2
anchorpy==0.18.0
